testing: remove dependency on kcl
We previously relied on kcl to create topics, delete topics, and delete
groups. There was no reason for that.

We now use kmsg directly, which simplifies the tests and removes some
cryptic depdendency. For now, we default to 127.0.0.1:9092 for seed
brokers, but that can be overridden with KGO_SEEDS. We may add cert/sasl
support later if necessary.