client: create a sink & source for all partition replicas
Previously, if using preferred read replicas, if a replica is not the
leader for any partition, then the sink&source is never created. The
leader will say "consume from 3", and the client will not have 3 saved
because it only saves the leader.

We now create a sink&source for all replicas of a partition as well.
Now, when a leader says "consume from 3", we should have 3 saved because
it will be in the Replicas portion of a MetadataResponseTopicPartition.

This also tidies up some repeated locking code for sinks and sources.