source: rotate through topics/partitions as we fetch
We previously rotated through partitions as we evaluated whether they
should be added to fetch requests, but this was meaningless because the
partitions are added to a two-layer map, which has random, mostly
not-actually-random range ordering.

We now keep our evaluated cursors in order, meaning we do eventually
rotate through partitions in order. Per topic, the partitions rotate
until we get have one full rotation, and then we rotate topics. Around
it goes.