metadata: allow leader epoch rewinds after 5 tries
See the embedded comments: if the entire cluster is in a bad state and
lost a leader epoch update, we should allow the client to recover. If we
"recover" into in accidental rewind, other aspects of the client will be
notified (produce may see NOT_LEADER_FOR_PARTITION, consume will see
FENCED_LEADER_EPOCH).

Closes #119.