txns: sleep 200ms on commit, preventing rebalance / new commit
See the updated transaction docs. This sleep will not happen if the user
requires stable fetch offsets and the cluster is 2.5+.

Ideally 200ms is fast enough especially in the context of this
preventing and incredibly rare scenario. However, the scenario is such
that if it happens, unknown undetectable duplicates could occur. 200ms
may not be long enough in light of this, but the thought here is to
allow the cluster at least a little bit of a heads up before allowing a
rebalance, which itself requires JoinGroup and SyncGroup to complete
before OffsetFetch can be issued to cause this problem.