config: add duplicately specified topic validation
Specifying both ConsumePartitions and ConsumeTopics is valid because
this allows a union of some partitions from one topic and all of a
second topic. However, it is invalid to try to consume some of topic
"foo" while also specifying all of "foo" in ConsumeTopics. Previously,
we would just consume all of "foo" and ignore the partial specification.
Now, we error if the topic is specified in both locations. Ideally, this
reduces some confusion.