kgo: superlint
This is most of f739622, but backs out of some of the lints I don't
really care for, actually. In particular, some cases of switching on the
return for kerr.ErrorForCode -- this type can *only* be unwrapped kerr
errors, so using errors.Is is overkill and makes the code look awkward.
As well, I want to use nil contexts in some scenarios, deliberately
internally.

From the original linting commit:

After tests found a race, I decided to super lint this package. I have
occasionally ran staticcheck against this and kept the bits I disagree
with, but it's easy enough to convert to a more "consistent" style and
nolint the ones I still disagree with.

The logic after this commit should be logically identical to what it was
before. The one bug found is going to be in a separate dedicated
commit, as will be the golangci-lint configuration.