config: add UnknownTopicRetries option, use more widely
If a topic was produced to successfully but then subsequently deleted,
the client would forever try sending to that topic. We repeatedly
receive UNKNOWN_TOPIC_OR_PARTITION errors.

In realistic Kafka environments, this error is only returned when things
do not exist. We should not try to forever send to a broker when the
cluster is telling us the topic is gone.

We previously failed all records for unknown topics if the client never
loaded the topic to begin with; we now extend that logic for if the
topic becomes unknown while producing.

We also lower the default fail limit from 5 to 4 (which in combo with
the prior metadata min age of 2.5s change, means we will detect the
problem and fail within 10s), and we add an option to configure this
number.