consuming: handle exact offset consuming better
Previously, if an exact offset was before the start offset or after the
end, the client would loop on OffsetOutOfRange errors. Now, we validate
the bounds and reset either to the start offset or the end offset if the
exact offsetis too low or too high, respectively.

We also better document this behavior.