client: cache sharded admin metadata lookups
Previously, if a sharded request needed to know where to split
partitions to, the client would *always* issue a metadata request to
load the freshest location of where to go. This was easier than
handling the response and splitting it, etc.

Upon further thought, splitting a partially failed response is
unnecessary. Instead, we can just retry the request entirely. The
non-retryable portions will just fail again, but our cache will be
updated such that we can send things to the more proper location.

Now, we add caching: on failure, if we ever see unknown topic or
partition, we delete the entire topic from our cache.

Because the user can out of band add partitions to a topic, we cannot
just rely on cached data forever. To start, we just use the metadata min
wait which by default is not much but does help the extremely common use
case of loading start & end offsets back to back. We may introduce
another config option to have a longer admin metadata wait, since only
CreatePartitions can break the cache. We also specifically hook into
CreatePartitions to evict our cache.

We also use simpler errors for when partitions are not returned from
metadata lookups, which also allows us to hook into the simpler errors
with the logic above and refresh our cache if necessary.

This also removes some confusing documentation from MetadataMinAge
(looking at the function since it is relevant for the above), and
ensures that topics can be purged from the mappedMeta map with
PurgeTopicsFromClient.