consumer group: track topics that the leader is not interested in
Previously, the group leader would not track topics that it is not
interested in. If A joins interested in topic "foo", and B joins in
topic "bar", if A is leader, it would balance foo and bar correctly, but
then it would not watch for metadata updates on bar.

Now, if "external" topics are detected, the group member stores all
group topics. These are used in metadata requests and the external
topics are now checked for changes after metadata updates.