producing: doubly use r.Offset for other numbers
Previously when buffering records, we would use a dedicated struct
promisedNumberedRecords, to keep track of the record length field and
the timestamp delta.

We now stuff these two int32's into the record Offset field, since the
Offset field is only set after a record is produced.

We also document that the offset will will not be valid unless the
record was successfully produced.