producer: serialize promises
Having promises potentially be concurrent was originally thought to be
an optimization, but overall after extensive benchmarking, it may not be
as much of an optimization as thought.

Previously, records that failed independent of producing could have
their promises called whenever. Produced records would always have their
promises called at the end of handling a produce response. Technically,
this means that if a broker has max produce requests in flight, starting
a new produce request would block on finishing promises for a previous
one if the promises were slow.

Now, we send all promises to a dedicated loop. This keeps our prior
guarantee of strict ordering and divorces promises from produce
requests.

I expect the common case for promises is to serialize the records
anyway, which implies locking within the promise. Hopefully, that
locking can now be removed. It is possible that promises that did no
work or did atomic work could be slower now, but extensive testing with
franz-go's bench utility (and changing the code to auto-success produce
requests) shows no difference -- if anything, serializing may be
slightly faster.