client: add EndAndBeginTransaction
For situations where you just want to end a transaction and begin a new
one, without caring about what *exactly* is flushed, it can be
beneficial to allow concurrent producing while end / begin are ongoing.
In particular, the restriction to flush may add unexpected latency.

This adds EndAndBeginTransaction that can be called concurrent with
producing. There are two safety options to choose: one that does not
really increase throughput **that** much, but does keep things safe,
and one that relaxes throughput at the expense of safety.

We delete some logic from txn_test to use our new function. Previously,
we would know what all buffered records are flushed when ending a
transaction. That is no longer true, so we can no longer count on the
transactional marker being where we expect. Instead, we ensure our
offsets are at least monotonically increasing and call it good.