producer: allow a canceled context & aborting to quit unknown wait
Unknown topic waiting was missing two areas to "quit" that existed
_after_ a topic was loaded (and no longer unknown). Aborting or a
canceled record context would not quit waiting.

We now quit on both. For aborting, we add a new channel to the unknown
type for must-quit errors. Previously, all errors being sent to
unknown.wait would be retriable (either retriable kerr's, or retriable
client connection failures). We do not want to send our must-quit signal
to the unknown.wait channel because *theoretically* it could be fully
buffered and thus either we would have to (a) wait to send, which has
problems if the wait loop quits, or (b) select/default, which would drop
our must-quit error.

We now use a size-1 channel that, if ever received from, always quits
the loop. As well, for a ctx-canceled record, similar to batches, we use
the first record's context.

Closes #158.