add OnOffsetsFetched function to allow inspecting commit metadata
This can be used to inspect the metadata of offsets that were assigned.
Offset fetching happens after assignment, meaning we cannot just stuff
this response into the context of onAssigned. As well, we want
heartbeating to be concurrent with offset fetching, and we want
heartbeating to begin after the user onAssigned callback because
heartbeating can also notice and trigger an onRevoked.

We do guarantee that OnOffsetsFetched (and now guarantee that
AdjustOffsetsBeforeAssign) will not be concurrent with
OnPartitionsRevoked, though.