BalancePlan: add AsMemberIDMap
The BalancePlan is opaque so as to reserve modification to its internal
structure. This modification is not likely, and having visibility into
the current plan may be beneficial.

We now add an API that returns the member => topic => partitions map,
and explicitly document that the underlying structure may change and
that the API will try to be as forward compatible as possible.