kadm: fix long standing poor API (Into fns)
The Into functions for converting OffsetResponses, OffsetsList, and
ListedOffsets into Offset, as well as Offsets into kgo.Offset, were
non-obvious functions.

We add better names, for nowdeprecating the old. When we address the
rest of #141, we will remove the deprecated functions before stabilizing
the package.

The rest of the API seems pretty okay-ish for how large it is.
NewClient/NewOptClient is a little awkard, but any alternative names
aren't much better (ClientFromKgo/NewClient is my preference, but...
this is now widely used).