kgo: improve ConsumeResetOffset, NoResetOffset, add Offset.AfterMilli
ConsumeResetOffset was confusing, and I hope the added docs on it make
it much clearer.

As well, NoResetOffset was limited, had some flaws, and required
fetching from the start. We now change the internals of Offset so that
NoResetOffset can be used at any specific offset (start, end, exact,
etC).

Lastly, this adds AfterMilli, which covers the most useful parts of
KAFKA-12478. For us, AfterMilli can be used in tandem with
NoResetOffset, meaning the user can opt to reset properly or not.
"latest-on-start" basically *is* AfterMilli, so the additional timestamp
aspects don't seem to help.

The "nearest" bit brought up is already done automatically, and is
arguably the better behavior. If needed someday, we can opt out of
automatic nearest bounding. This was the default pre 1.3.5, and tripped
people up occasionally.

"earliest-on-start" seems useless, because the implication is that we
want earliest on start OTHERWISE latest, and I can't think of a use case
for that. If we don't use OTHERWISE latest, then this is just
"earliest".

Closes #154.