add package sr, providing a client for the schema registry
This punts responsibility of encoding & decoding functions to the user,
but the Client and Serde types should make it easy to at least use a
schema registry and encode/decode things properly.

This supports the entire API except exporters, which is currently in
preview. Some confusing schema registry functions are abstracted to be a
bit more obvious -- rather than returning a bunch of IDs, we go one step
further and map those IDs back to the SubjectSchema types.

I'm keeping this as a separate module for now so that I can actually
play around with the API.