kgo: improve consuming/producing *immediately* after topic creation
If you were consuming immediately after topic creation, the client could
first see UNKNOWN_TOPIC_OR_PARTITION and then backoff 2.5s. This is
because when initializing a client, we used triggerUpdateMetadata.
It doesn't hurt much to switch to triggerUpdateMetadataNow for client
init, and beter supports the (rare) use case of consuming immediately
after client init.

For producing, Kafka can respond to metadata that a partition has a
leader before the leader actually *is* a leader of the partition. If we
produced immediately after topic creation, we could sometimes backoff
2.5s. We now switch to behavior where if this is the *first* record
being produced to a partition (per the record seq, which yes can be
reset in rare cases but doesn't really hurt this fix), and if the error
is due to UNKNOWN_TOPIC_OR_PARTITION, rather than triggering a metadata
update and waiting, we set the partition to backoff temporarily and
retry.