producer: allow the user to set Timestamp
I'm not sure why I introduced the restriction that timestamps must be
sequential. It appears the MaxTimestamp field just needs to be the max
timestamp of the batch, and FirstTimestamp (BaseTimestamp) just needs to
be the timestamp of the first record. If timestamps go backwards,
nothing in Kafka seems to care / check this.

Tested by producing two records with a linger where the second record
has a 1s earlier timestamp than the first record.