kgo: add optional interface to close sasls on client close
This will be used in the kerberos module once v1.7 of franz-go is
released soon. As seen in #182, the inability to close the kerberos
mechanism makes it really easy to leak goroutines. There are staged
changes to help make using kerberos more foolproof, but they are
dependent on this commit: check all sasl mechanism to optionally close
them on shutdown.