kgo client.go: bugfix -1 leader epoch for a few minutes
When the previous brokers map was split into brokers and seeds fields,
the supportsOffsetForLeaderEpoch was not updated to check seeds if no
brokers existed yet. This means that for the first metadata update, we
would store -1 for the leader epoch, and down the line when metadata
updated again, we would store the real leader epoch. The fallout is a
bit nebulous. I think this means we have a small window of time where we
would not validate data loss, but that's about it.