README: expand installation section more
This adds a small block for the go.mod bit, and describes how kgo
relates to the repo.

It is common for repos to provide one package, and thus the repo root is
the package. It's not always the rule, though: repos with go-xyz names
may have a go-xyz/xyz package, and large repos (aws-sdk-go) have dozens
of packages.

Closes #183.