kgo.Fetches: add Err0
I'm having a hard time thinking of a good name here, and at least Err0
somewhat implies the indexing aspect.

I've tried:

* QuitPollErr, which too focused because it would imply that it only
returns non-nil if the error is context.Canceled or ErrClientClosed,
i.e. I would need to check the error.

* SingleErr, which is a bit awkward.

* OnlyErr, similarly awkward (and what is proposed in #189).

After staring at 189, it is a bit compelling to have very fast closed /
canceled checks, especially if the user consumes a LOT of partitions.

Credit to @dwagin for the idea / initial implementation.

Closes #189.