kgo: add Context field to Record and set while producing
See #201 and #200.

Some packages want to stuff information into a context with WithValue,
and then that information can be used to trace spans (see the open
telemetry packages). Without a Context field or something equivalent, it
is not possible to trace spans once a record is handed off to franz-go.
We want to use these spans in franz-go hooks.

We do not need to do anything on the consumer side, but unfortunately,
we need the Context field to be **writable** so that consumers can
initialize the context (and stuff information into it) with WithValue.
So, we just add a new field to the record.