UniformBytesPartitioner: note sub-optimal batches with lingering
This is harder / impossible to fix in the same way as it is in Kafka:
the wiring for lingering is deeper than the partition pick, so picking a
partition has no context of lingering. We cannot take into account
lingering still trying to fill a batch.

The fix in Kafka *looks* like it is not perfect in the context of
disabling lingering. In all, there's probably no golden fix for all
approaches.

We'll just note the problem, since this client recommends disabling
lingering anyway.