RecordFormatter: support %a; formatter&reader: support 'bool'
Attributes are special in that there is no obvious way to print them:
a single record has many attributes that all mean different things.

We instead add a required formatting specification that selects which
attribute to print. The specification can have ; to separate _how_ to
print the attribute.

We also support formatting booleans as 'true' or 'false', and reading
'true' as 1 and 'false' as 0. The reading is a bit complicated and uses
some weird closure. The code is ..a bit ugly.. but it is fun and tested.