kadm: ListOffsetsAfterMill(future) should return end offsets
Previously, ListOffsetsAfterMilli(future ts) would return -1 offsets for
timestamps in the future / for when partitions do not have a record
after the requested timestamp.

After using this for a while, IMO this is not desirable behavior: if I
don't have an offset after a given time, then I want the end offset.
This changes the behavior to match that.