kgo: add ErrFirstReadEOF, which unwraps to io.EOF
The client previously would sometimes fail requests with io.EOF, and
internally it would log "perhaps sasl?". We now return an error that
indicates the same thing. This can be used by end users to check and
potentially warn themselves if sasl may be missing. Because this unwraps
into io.EOF, we do not need to change any "is this error retryable"
functions internally within kgo.