kgo: add DialTimeout function, complementing DialTLSConfig
Using a non-10s dial timeout is "easy" but more cumbersome than it
should be. It's also not so easy if the user wants to use a TLS
configuration.

Adding DialTimeout allows the user to just specify a tls config and a
timeout and have the client dial as they expect, rather than requiring
the user to re-implement DialTLSConfig just to change the timeout.