kgo: support forward & backward batch requests for FindCoordinator, OffsetFetch
Previously, the client was only forward compatible. Opting into batching
was at your own risk. We now default to batching and split to the old
single behavior as needed.

To support this, we have a new internal request version pinner. We
always pin to batched version, and if we get errBrokerTooOld, we split.

loadCoordinators is a good bit more complex now, and we basically delete
loadCoordinator.

This is a bit complicated to describe in a single commit message.