{Cooperative,Sticky}Balancer: bug fix lack of stickiness
We clear nowAssigned at the end of a group session, meaning using it for
our prior assignment is invalid and would always be empty. Thus, we
would never preserve stickiness.

Now, we use our lastAssigned for stickiness, we *always* set it, and we
set it after using the current/prior values and before we clear
nowAssigned.