kadm: bugfix DeleteRecords,DescribeProducers,OffsetForLeaderEpoch
In sharded requests, if we range over shards and set values in a map, we
could overwrite values we created on a previous shard. This commit
adds comments or fixes any incorrect shard handling.

Thankfully, two problems were in recently introduced APIs (<1w).
Unfortunately, one problem is in DeleteRecords, an old API.