source: avoid backoff / session reset when there is no consumed data
We use fetch sessions by default. 2b4d7ea8 in 1.9.0 fixed a potential
spin-loop, but introduced a different minor bug: if broker returned
nothing to consume for our fetch session, the broker would return no
partitions and we would (a) backoff, (b) trigger a metadata update, and
(c) reset our fetch session.

The fix changed the source to back off if all partitions were stripped
due to retriable errors, but our check was against the number of
partitions in the response. We now check against the number of
partitions in the request (numOffsets).