sink: tighten canFailFromLoadErrs
The old version set canFailFromLoadErrs=false immediately once a batch
was added to a request, well before the request is written. It is
possible to have failures and delays that cause the request to never be
written, but our record could never be failed.

We now only set canFailFromLoadErrs=false at the moment we are about to
write the request.

Closes #239.