kgo: allow empty groups when finding coordinator / fetching offsets
v1.10.0 introduced a bug when switching to batch loadCoordinators:
we would strip empty coordinator keys from the request.

Technically, empty coordinator keys are valid. If a request contains an
empty coordinator key (empty group), we would not load it and not set a
map key for it and then using a field in the map value would panic.

Now, we always request all coordinator keys.

This also showed a bug in sharded OffsetFetch: fetching offsets for a
group with no name would be stripped (now we will just forward whatever
error happens).

Closes #283.