kgo: fix batch coordinator fetching
v1.10.0 introduced batch coordinator fetching, but had a bug: when
fetching many coordinators in a batch, we would not use the proper
NodeID (we would use the top level NodeID, not the per-coordinator
NodeID).

This worked for non-batch coordinator fetching because we mirrored the
first coordinator to the top level response, if the request/response was
batched.