producer: set Record.Topic earlier
Certain applications may need to know the record topic in
OnRecordBuffered, but if the application used DefaultProduceTopic and
then did not bubble that default topic through the entire app, knowing
the record's topic was not possible.

Now, we set the record topic earlier.

We also now pre-emptively fail records before hanging waiting for space
if we know the record is immediately invalid, rather than waiting for
buffer space and then failing.